Enhance Docker registry configuration and remove unused module files
This commit adds a lifecycle block to the Docker registry configuration to ignore changes to metadata labels. It also updates the Kyverno policy to synchronize secrets with the Docker registry. Additionally, it removes obsolete module.mk files for ArgoCD and OLM, streamlining the project structure. The paths for OLM artifacts in the main.tf file have been corrected to ensure proper resource referencing.